Cleanup stdout/stderr polling
Prior to this commit we were using a while loop to query stdout/stderr
returned by a running process and we had two separate exception
handlers for any `WaitReadable` exceptions that occured.
This commit removes the while loops because from what I can tell
they are unnecessary, as well as consolidates the `WaitReadable`
rescue blocks to one area that will print that the process
is waiting to `stderr`.