Add FilePush
Prior to this commit if you wanted to upload a file and copy it
to a location, you needed to make a temp dir, copy the file to that
temp dir on the host, move the file to its final destination, chmod
and chown the file, and cleanup the tempdir.
This commit adds the `FilePush` action that performs all of that
busywork for you.