(PE-25571) CVE-2018-6517 Stop automatic known_hosts additions
Prior to this commit, Chloride used a version of the net-ssh gem that
would automatically add an entry to the user's known_hosts file when
connecting to an unknown host. This is considered a security risk
(CVE-2018-6517).

This commit updates the net-ssh gem requirement to version 4, which has
added custom known_hosts file support. I have implemented
`Chloride::SSHKnownHosts`, which will create a separate, library
specific, known_hosts file in the current directory where new SSH host
signatures that were not found in the system known_hosts will be added.
This file path is configurable via the :known_hosts option when
instantiating the `Chloride::Host` object.